services/api/pyproject.toml
[project]
name = "historicam-api"
version = "0.1.0"
description = "HistoriCam building identification API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "google-cloud-aiplatform>=1.70.0",
    "vertexai>=1.43.0",
    "pillow>=11.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
